requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
charset-normalizer==2.0.3
Cython==0.29.24
distlib==0.3.2
Django==3.2.5
django-rest-framework==0.1.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
filelock==3.0.12
idna==3.2
numpy==1.21.1
openpyxl==3.0.7
pandas==1.3.1
platformdirs==2.0.2
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
virtualenv==20.6.0
